Using long for storing counts, to get the large dataset working

+++ Controllers/QuotesController.cs
@@ -27,7 +27,7 @@ namespace CodeChallenge.Controllers
 
         // GET api/quotes/pairs/20
         [HttpGet("pairs/{length}")]
-        public int GetPairs(int length)
+        public long GetPairs(long length)
         {
             return quoteRepository.GetNumberOfPairs(length);
         }
